+++ gamevault/ViewModels/GameDownloadViewModel.cs
@@ -124,7 +124,7 @@ namespace gamevault.ViewModels
         {
             get
             {
-                return new string[] { ".7z", ".xz", ".bz2", ".gz", ".tar", ".zip", ".wim", ".ar", ".arj", ".cab", ".chm", ".cpio", ".cramfs", ".dmg", ".ext", ".fat", ".gpt", ".hfs", ".ihex", ".iso", ".lzh", ".lzma", ".mbr", ".msi", ".nsis", ".ntfs", ".qcow2", ".rar", ".rpm", ".squashfs", ".udf", ".uefi", ".vdi", ".vhd", ".vmdk", ".wim", ".xar", ".z" };
+                return Globals.SupportedArchives;
             }
         }
         public Dictionary<GameType, string?> GameTypes